requirements.txt
pillow==11.3.0
python-dotenv
langchain-core
langchain-openai
pytest
transformers
PyYAML
safetensors
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
nltk
pandas